Add Global component and CSS variable
