+++ code/src/index.css
@@ -4,6 +4,8 @@
   --paragraph-font-family: "Newsreader", serif;
   --drop-shadow: drop-shadow(2px 4px 1px rgb(0 0 0 / 0.3));
   --box-shadow: 0 5px 10px rgba(0, 0, 0, 0.3);
+  --primary-background-color: #94b1aa;
+  --btn1-color: #94b1aa;
 
 }
